requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.3.0
babel==2.16.0
backoff==2.2.1
beautifulsoup4==4.12.3
bert-score==0.3.13
bleach==6.2.0
blinker==1.9.0
CacheControl==0.14.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.2.0
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
effdet==0.4.1
einops==0.8.0
emoji==2.14.0
et_xmlfile==2.0.0
eval_type_backport==0.2.2
evaluate==0.4.3
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.16.1
filetype==1.2.0
fire==0.7.0
firebase-admin==6.6.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.12.23
fonttools==4.55.3
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.9.0
gcloud==0.18.3
google==3.0.0
google-api-core==2.24.0
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
google-cloud-vision==3.9.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
gradio==5.12.0
gradio_client==1.5.4
groq==0.13.1
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
humanfriendly==10.0
idna==3.10
iopath==0.1.10
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.38
jaxlib==0.4.38
jedi==0.19.2
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-groq==0.2.3
langchain-huggingface==0.1.2
langchain-openai==0.3.0
langchain-pinecone==0.1.2
langchain-text-splitters==0.3.5
langdetect==1.0.9
langsmith==0.1.147
layoutparser==0.3.4
# llmsherpa @ file:///Users/numair/Desktop/numair_fuzzynodes/llmsherpa
llvmlite==0.43.0
lxml==5.3.0
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.25.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.20
mistune==3.0.2
ml_dtypes==0.5.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
ndjson==0.3.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.3.1
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
oauth2client==4.1.3
olefile==0.47
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.20.1
openai==1.58.1
openai-whisper==20240930
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opt_einsum==3.4.0
orjson==3.10.13
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pdf2docx==0.5.8
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pi_heif==0.21.0
pikepdf==9.5.1
pillow==11.0.0
pinecone==5.4.2
pinecone-client==4.1.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
portalocker==3.1.1
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
protoc-gen-openapiv2==0.0.1
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.10.1
PyMuPDF==1.25.1
pypandoc==1.15
pyparsing==3.2.0
pypdf==5.1.0
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.10.22
python-json-logger==3.2.1
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.1
python-pptx==1.0.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
ragas==0.2.11
RapidFuzz==3.11.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
ruff==0.9.2
safehttpx==0.1.6
safetensors==0.5.0
scikit-learn==1.6.0
scipy==1.14.1
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==3.3.1
sentencepiece==0.2.0
sh==2.2.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.6
SQLAlchemy==2.0.36
srt==3.5.3
stack-data==0.6.3
starlette==0.41.3
style==1.1.0
sympy==1.13.1
tenacity==8.5.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.8.0
timm==1.0.13
tinycss2==1.4.0
tokenizers==0.21.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.6.0.dev20241228
torchaudio==2.6.0.dev20241228
torchvision==0.22.0.dev20241228
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
twilio==9.4.3
typer==0.15.1
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
unstructured==0.16.12
unstructured-client==0.28.1
unstructured-inference==0.8.1
unstructured.pytesseract==0.3.13
update==0.0.1
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
vosk==0.3.44
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wrapt==1.17.1
xlrd==2.0.1
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.18.3
